% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{browsershares}
\alias{browsershares}
\title{Browser shares}
\format{An object of class \code{data.frame} with 11 rows and 9 columns.}
\source{
\url{https://en.wikipedia.org/wiki/Usage_share_of_web_browsers}
}
\usage{
browsershares
}
\description{
Global desktop and mobile.
}
\keyword{datasets}

