% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_get}
\alias{list_get}
\title{Extracts elements from a list based on index}
\usage{
list_get(l, ext, ...)
}
\arguments{
\item{l}{the list}

\item{ext}{the index}

\item{...}{addtional arguments}
}
\description{
Extracts elements from a list based on index
}

