% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_converter.R
\name{multicol_header}
\alias{multicol_header}
\title{Create multicolumn TeX headers}
\usage{
multicol_header(col.vars, offset.left)
}
\arguments{
\item{col.vars}{the column vars}

\item{offset.left}{offset from the rows variable}
}
\description{
Create multicolumn TeX headers
}

