% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_converter.R
\name{print.tischout}
\alias{print.tischout}
\title{Printing method for tisch code}
\usage{
\method{print}{tischout}(x, ...)
}
\arguments{
\item{str}{the text to be printed}
}
\description{
Printing method for tisch code
}

