% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisch.R, R/utils.R
\name{reorder_df}
\alias{reorder_df}
\title{Reorder the df}
\usage{
reorder_df(df, vec)

reorder_df(df, vec)
}
\arguments{
\item{df}{the data frame}

\item{vec}{the vector by which the data frame is sorted}

\item{row.names}{the row names}

\item{row.order}{the row order}

\item{col.order}{the col.order}

\item{df}{the data frame}
}
\description{
Reorder the df

Reorder a data frame
}

