% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fun.R
\name{spanned}
\alias{spanned}
\title{Create a spanend row style}
\usage{
spanned(h.offset = 0.1)
}
\arguments{
\item{h.offset}{numerical, the vertical offset between grouped rows in cm}
}
\description{
Create a spanend row style
}

