% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tisch.R
\name{tisch}
\alias{tisch}
\title{Create a new tisch object
tisch initializes a new tisch object. Designate rows and columns via its formula interface.}
\usage{
tisch(data, formula, reorder = FALSE, ...)
}
\arguments{
\item{data}{a data frame}

\item{formula}{\link[reshape2]{cast}}

\item{reorder}{logical, allow the casting function to reorder the data frame?}

\item{...}{arguments passed to the casting function}
}
\description{
Create a new tisch object
tisch initializes a new tisch object. Designate rows and columns via its formula interface.
}
\examples{
\dontrun{
data(asd)
tt <- tisch(data, asd + asd ~ asd)
to_tex(tt)

tt + caption("asd") + footnote("ASd") +
theme(asd)
to_tex(tt)

}
}

