% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_converter.R
\name{to_html}
\alias{to_html}
\title{Converts a tischbase object to HTML output}
\usage{
to_html(obj, ...)
}
\arguments{
\item{obj}{the tischbase object}
}
\description{
Converts a tischbase object to HTML output
}

