% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_converter.R
\name{to_tex}
\alias{to_tex}
\title{Converts a tischbase object to LaTeX output}
\usage{
to_tex(obj, ...)
}
\arguments{
\item{obj}{the tischbase object}
}
\description{
Converts a tischbase object to LaTeX output
}

