% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fun.R
\name{spanned}
\alias{spanned}
\title{Create a spanend row style}
\usage{
spanned(which = 1, ...)
}
\arguments{
\item{which}{asd}

\item{...}{additional arguments passed to nowehere}
}
\description{
Create a spanend row style
}

