% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_tisch}
\alias{write_tisch}
\title{Write the tisch object to the file system}
\usage{
write_tisch(x, filename)
}
\arguments{
\item{x}{tisch object}

\item{filename}{a file path string}
}
\description{
Write the tisch object to the file system
}

